ncs 469 moved endpoint to its own controller
